End of Mon Jul 8 work hour

+++ include/blob_extractor.hpp
@@ -12,7 +12,7 @@ public:
   // float search_radius: how far should be considered connected
   // int [min|max]_volumn: what size should be considered a blot
   ImageBlobExtractor(float focal_length, float search_radius,
-		     int min_volumn, int max_volumn,
+		     float min_volumn, float  max_volumn,
 		     int min_count,  int max_count);
   
   void setInputCloud(CloudPtr cloud);
@@ -23,7 +23,7 @@ public:
 private:
   const float f;         // focal length
   const float search_r;  // search radius
-  int min_v, max_v;      // [min|max] volumn
+  float min_v, max_v;    // [min|max] volumn
   int min_c, max_c;      // [min|max] point count
   CloudPtr cloud;
 };
